Lecture 43: intro to CSS
It is possible to format and style html elements from the .html file
itself, by adding attributes to those elements. However, we don't want
to do that, and want to separate the syles we use in another .css file.

To link the two files, the .html and the .css files, we use a <link> tag
inside the header.

All elements will be affected by what is in the .css file.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>