lecture 48: styling elements via 'class'
If we want to style many items at once, we can use the 'class' attribute
on that element.

The same class can be assigned to many elements, making a single css
change affect many elements on a page.

The class attribute must start with a dot '.' before its name in the
.css file.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>