lecture 52: order of hierarchy
The CSS inheritance is calculated by assigning numeric values to each
of the elements and element combinations. The hierarchy is described in this
commit.
Type selectors are the least specific;
Then comes classes, attributes, and pseudo-class selectors;
And the most specific is ID's.

Reference: https://code.tutsplus.com/tutorials/the-30-css-selectors-you-must-memorize--net-16048

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>