lecture 61: the box model
Every CSS element has a conceptual box sorrounding it, consisted of:

    - margin;
    - padding;
    - border;
    - content;

Margin is the space between the border and the next element in the page.
Padding is the space between the border and the content, i.e. the inner
side of the box.
Border is the line sorrounding the box.
And content is the text, or image, etc, of the box.

All four properties have four sides: Top, Right, Bottom, and Left, and
they can be set or customized in this order. When setting a property,
keep in mind that, for instance:

    margin: 20px 40px 500px 100px;

refers to
    margin: <top> <right> <bottom> <left>;

And when we see:

    margin: 0 auto;

it means:

    margin: <top+bottom> <right+left>;

'auto' above means an automatic calculation made by the browser
according to the size of the parent, which is the 'body' in this code's
case.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>