lecture 76: add jumbotron and container
Jumbotron is a bootstrap element consisting of a box to highlight some
contents of you site. In this jumbotron there is an h1, a paragraph and
a button.

Also a div with a class "container" was introduced in order to group
contents in a box with margins and some predefined styles.

Find the jumbotron documentation of jumbotron under Components at [0].

[0] https://getbootstrap.com/docs/3.3/components/#jumbotron

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>