lecture 76: add forms
Add 4 different types of forms: email, password, file and submit.

Important notes are the .form-group and .form-control classes for the
div and input element, respectively.

The form-group class collects all elements inside the div as a group.
The form-control class provides some styles from bootstrap.

The type="file" in an input element will indicate that a file should be
selected from the filesystem and will open the default file manager so
that the user could select it.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>