lecture 77: add navbars
Add a default navbar including links on the left and on the right. Also
make it responsive to the screen size. When shrinked down, the navbar will
hide selected contents and group them in a button that will toggle them
when it is pressed.

Two dependencies were imported: JavaScript and JQuery, taken from [0]
and [1] respectively, and embedded into the code as suggested by the web
pages.

[0] https://getbootstrap.com/docs/3.3/getting-started/#download
[1] https://code.jquery.com/

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>