lecture 80: the grid system - part II
Fine tune the number of items we want to show in each row, depending on
the size of the screen.

The classes determine the quantity: md-3 indicates that for the medium
size of the screen, each item should take up 3 blocks out of the 12,
which will make 4 blocks per row.

Likewise, sm-6 will make it render 2 block per row when the screen size
reaches the small size (because 12/6 gives 2 blocks per row).

This behavior can be seen in the tour dates section on the url [0].

[0] http://olafurarnalds.com/

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>