lecture 82: add grid system for photos
The grid system is implemented. Note the use of a new bootstrap
class 'thumbnail'. The thumbnail class creates a div that sorrounds the
image and shrinks it to the size of the column.

Regarding the columns, the column classes were specified as:
    col-lg-4: each colum will take 4 blocks out of 12 total, making a row
	      with 3 images.
    col-sm-6: each column will take 6 blocks, making rows with 2 images
              each.

The medium 'md' size is not necessary because bootstrap interprets it as
the same size as the next smaller one (in this case, 'sm').

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>