lecture 84: change icon and styles
Overwrite styles coming from bootstrap by customizing them on the .css file.
Note that for that to work, the .css import needs to be after the bootstrap
one, for it to be loaded last and override the value from bootstrap.

Also we change the glyphicon used in the jumbotron by one from fontawesome,
which was also imported in the header section.

Lastly, change two photos just for experimenting. Note that some tweaks were
done in the first image so that the size of the photo would fit the block.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>