lecture 85: style fonts and horizontal rule
Add the Lato font family and make them all white. The button will and the
navbar will not be affected by the white color because they have more
specificity provided by bootstripe.

The h1 font size (5em) means that its size should be 5 times bigger than
the default size (defined in the body - which in this case is the HTML default
size. From the Mozilla MDN page [0]:

"The size of an em value is dynamic. When defining the font-size property, an
em is equal to the size of the font that applies to the parent of the
element in question."

[0] https://developer.mozilla.org/en-US/docs/Web/CSS/font-size

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>