lecture 86: add icons and font shadows
Again, use fontawesome to get some cool icons (paw, login and sign-up).
Note that fontawesome is providing the CND link upon subscription, and the
link used here is the provided in the course, on lecture 84.

Also note that the "Get Started" button was being under influence of font
changes made in class #content because it was not a 'button' element (it was
an 'a' element). This commit also fixes it.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>