lecture 108: add background and with navbar
This new project is called Museum Of Candy. This is a code along project
and is provided as a resource in two forms the course instructor:
a Starter and a Final code for comparison/reference.

The navbar was added by hand based on the code provided by
Bootstrap 4 documentation page [0], but not copying and pasting it, so some
contents may differ from the one provided in the web page.

Disclaimer: I will code along and some snippets may differ from the Final code.

[0] https://getbootstrap.com/docs/4.3/components/navbar

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>