lecture 109: add media query to reduce text size
This commit introduces a media query to manipulate the formatting of the
'#headingGroup h1' when the browser reaches a certain size (in our case,
1200px and 900px).

The @media rule includes a block of CSS properties that change when a
certain condition is true (in our case, the browser width: 'max-width').

Inside the @media block we repeat the '#headingGroup h1' specificity to
decrease its size in two distinct sizes, before it disappears when the
screen size reaches the small(s) size (implemented in commit #c1238de6).

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>