lecture 111: add two more divs to the main page
This commit duplicates the first div to add two more divs with images and text.
Note that the second div has the text on the left while the image on the
right.

Also some minor adjustments were made to the borders so the layout is a little
cleaner on smaller browser sizes.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>